+++ TrabExtClasse/.gitignore
@@ -1,0 +1,2 @@
+/nbproject/private/
+/build/
