--- lib/nblibraries.properties
@@ -1,8 +1,0 @@
-libs.CopyLibs.classpath=\
-    ${base}/CopyLibs/org-netbeans-modules-java-j2seproject-copylibstask.jar
-libs.CopyLibs.displayName=Tarefa CopyLibs
-libs.CopyLibs.prop-version=2.0
-libs.MySQLDriver.classpath=\
-    ${base}/MySQLDriver/mysql-connector-java-5.1.23-bin.jar
-libs.MySQLDriver.displayName=Driver JDBC do MySQL
-libs.MySQLDriver.prop-maven-dependencies=mysql:mysql-connector-java:5.1.23:jar
